fix(artifacts): close guide cutover review findings
