+++ scripts/test_lightweight_agent_gates.py
@@ -56,6 +56,46 @@ class LightweightAgentGateTests(unittest.TestCase):
             failures,
         )
 
+    def test_stale_artifact_rejects_legacy_guide_durable_ref(self) -> None:
+        failures = scan_artifact_text(
+            "backend/app/modules/projects/example.py",
+            "Caller supplied durable_" + "ref.",
+            "guide_source_cutover",
+        )
+        self.assertIn(
+            "backend/app/modules/projects/example.py:1: LEGACY_GUIDE_DURABLE_REF",
+            failures,
+        )
+        interface_failures = scan_artifact_text(
+            "backend/app/interfaces/project_agents.py",
+            "Caller supplied durable_" + "ref.",
+            "guide_source_cutover",
+        )
+        self.assertIn(
+            "backend/app/interfaces/project_agents.py:1: LEGACY_GUIDE_DURABLE_REF",
+            interface_failures,
+        )
+
+    def test_stale_artifact_rejects_legacy_guide_content_hash(self) -> None:
+        failures = scan_artifact_text(
+            "backend/app/modules/projects/example.py",
+            "Caller supplied content_" + "hash.",
+            "guide_source_cutover",
+        )
+        self.assertIn(
+            "backend/app/modules/projects/example.py:1: LEGACY_GUIDE_CONTENT_HASH",
+            failures,
+        )
+        interface_failures = scan_artifact_text(
+            "backend/app/interfaces/project_agents.py",
+            "Caller supplied content_" + "hash.",
+            "guide_source_cutover",
+        )
+        self.assertIn(
+            "backend/app/interfaces/project_agents.py:1: LEGACY_GUIDE_CONTENT_HASH",
+            interface_failures,
+        )
+
     def test_stale_artifact_rejects_unknown_phase(self) -> None:
         with self.assertRaises(ValueError):
             phase_index("unknown")
